+++ src/main/java/com/swithus/community/club/controller/ClubController.java
@@ -1,8 +1,11 @@
 package com.swithus.community.club.controller;
 
+import com.swithus.community.club.dto.SearchPageRequestDTO;
+import com.swithus.community.club.service.ClubService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
@@ -11,10 +14,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequiredArgsConstructor
 @RequestMapping("/club")
 public class ClubController {
+    private final ClubService clubService;
 
     @GetMapping("/search")
-    public void goSearch() {
+    public void goSearch(SearchPageRequestDTO pageRequestDTO, Model model) {
         log.info("GET /club/search");
+
+        model.addAttribute("list", clubService.getSearchPage(pageRequestDTO));
     }
 
     @GetMapping("/main")
