+++ src/main/java/com/swithus/community/club/entity/ClubImage.java
@@ -1,19 +1,22 @@
 package com.swithus.community.club.entity;
 
 import com.swithus.community.global.entity.BaseEntity;
+import jakarta.persistence.*;
 import lombok.*;
 
-//@Entity
+@Entity
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-//@Table
+@Table
 @ToString
 public class ClubImage extends BaseEntity {
-
-    private Long key;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
     // 클럽 키
+    @ManyToOne(fetch = FetchType.LAZY)
     private Club club;
 
     // 이름
