+++ src/main/java/com/swithus/community/club/entity/ClubInquiryHeader.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 public class ClubInquiryHeader extends BaseEntity {
 
-    private Long key;
+    private Long id;
     // 클럽 키
     private Club club;
     // 회원 키
