+++ src/main/java/com/swithus/community/club/entity/GreetingsLike.java
@@ -4,7 +4,7 @@ import com.swithus.community.global.entity.BaseEntity;
 
 public class GreetingsLike extends BaseEntity {
 
-    private Long key;
+    private Long id;
     // 가입인사 키
     private Greetings greetings;
     // 누른 사람 키
