+++ src/main/java/com/swithus/community/club/entity/MeetingCtgr.java
@@ -4,7 +4,7 @@ import com.swithus.community.global.entity.BaseEntity;
 
 public class MeetingCtgr extends BaseEntity {
 
-    private Long key;
+    private Long id;
 
     // 모임 종류를 알려주는 카테고리명
     private String name;
