+++ src/main/java/com/swithus/community/club/repository/ClubImageRepository.java
@@ -1,0 +1,7 @@
+package com.swithus.community.club.repository;
+
+import com.swithus.community.club.entity.ClubImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ClubImageRepository extends JpaRepository<ClubImage, Long> {
+}
