--- src/main/java/com/swithus/community/club/repository/ClubLeaderRepository.java
@@ -1,4 +1,0 @@
-package com.swithus.community.club.repository;
-
-public interface ClubLeaderRepository {
-}
