+++ src/main/java/com/swithus/community/club/repository/search/ClubSearchRepositoryImpl.java
@@ -6,13 +6,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
 
 public class ClubSearchRepositoryImpl extends QuerydslRepositorySupport implements ClubSearchRepository {
+    // 생성자
     public ClubSearchRepositoryImpl() {
         super(Club.class);
     }
 
     @Override
     public Page<Object[]> clubSearchPage(Pageable pageable, int region, int sports, String keyword) {
-        // Qclub
 
         return null;
     }
