+++ src/main/java/com/swithus/community/club/service/ClubService.java
@@ -1,6 +1,79 @@
 package com.swithus.community.club.service;
 
-public interface ClubService {
+import com.swithus.community.club.dto.ClubDTO;
+import com.swithus.community.club.dto.SearchPageRequestDTO;
+import com.swithus.community.club.dto.SearchPageResultDTO;
+import com.swithus.community.club.entity.Club;
+import com.swithus.community.club.entity.ClubImage;
+import com.swithus.community.global.dto.ImageDTO;
+import com.swithus.community.global.entity.Region;
+import com.swithus.community.global.entity.Sports;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public interface ClubService {
+    SearchPageResultDTO<ClubDTO, Object[]> getSearchPage(SearchPageRequestDTO requestDTO);
+
+    default Map<String, Object> dtoToEntity(ClubDTO clubDTO) {
+        Map<String, Object> clubMap = new HashMap<>();
+
+        Club club = Club.builder()
+                .id(clubDTO.getClubId())
+                //.leader(User.builder().key(clubDTO.getLeaderKey()).build())
+                .region(Region.builder().id(clubDTO.getRegionId()).build())
+                .sports(Sports.builder().id(clubDTO.getSportsId()).build())
+                .name(clubDTO.getName())
+                .introduce(clubDTO.getIntroduce())
+                .rank(clubDTO.getRank())
+                .point(clubDTO.getPoint())
+                .build();
+
+        clubMap.put("club", club);
+
+        List<ImageDTO> imageDTOList = clubDTO.getImageDTOList();
+
+        if (imageDTOList != null && !imageDTOList.isEmpty()) {
+            List<ClubImage> imageList = imageDTOList.stream()
+                    .map(imageDTO -> ClubImage.builder()
+                            .path(imageDTO.getPath())
+                            .name(imageDTO.getName())
+                            .uuid(imageDTO.getUuid())
+                            .club(club)
+                            .build())
+                    .toList();
+
+            clubMap.put("imageList", imageList);
+        }
+
+        return clubMap;
+    }
+
+    default ClubDTO entityToDTO(Club club, List<ClubImage> imageList, Long personnel) {
+        List<ImageDTO> imageDTOList = imageList.stream()
+                .map(image -> ImageDTO.builder()
+                        .path(image.getPath())
+                        .name(image.getName())
+                        .uuid(image.getUuid())
+                        .build())
+                .toList();
+
+        return ClubDTO.builder()
+                .clubId(club.getId())
+                //.leaderKey(club.getLeader())
+                .regionId(club.getRegion().getId())
+                .regionName(club.getRegion().getName())
+                .sportsId(club.getSports().getId())
+                .sportsName(club.getSports().getName())
+                .name(club.getName())
+                .introduce(club.getIntroduce())
+                .personnel(personnel)
+                .rank(club.getRank())
+                .point(club.getPoint())
+                .regDate(club.getRegDate())
+                .imageDTOList(imageDTOList)
+                .build();
+    }
 
 }
